+++ src/com/landawn/abacus/util/Any.java
@@ -119,7 +119,7 @@ abstract class Any<T> {
      * otherwise do nothing.
      *
      * @param consumer block to be executed if a value is present
-     * @throws IllegalArgumentException {@code consumer} is null
+     * @throws IllegalArgumentException {@code action} is null
      */
     public <E extends Exception> void ifPresent(Try.Consumer<? super T, E> action) throws E {
         N.checkArgNotNull(action);
@@ -134,7 +134,7 @@ abstract class Any<T> {
     *
     * @param action
     * @param emptyAction
-     * @throws IllegalArgumentException {@code consumer} or {@code emptyAction} is null
+     * @throws IllegalArgumentException {@code action} or {@code emptyAction} is null
     */
     public <E extends Exception, E2 extends Exception> void ifPresentOrElse(Try.Consumer<? super T, E> action, Try.Runnable<E2> emptyAction) throws E, E2 {
         N.checkArgNotNull(action);
@@ -167,7 +167,7 @@ abstract class Any<T> {
     *
     * @param action
     * @param emptyAction
-     * @throws IllegalArgumentException {@code consumer} or {@code emptyAction} is null
+     * @throws IllegalArgumentException {@code action} or {@code emptyAction} is null
     */
     public <E extends Exception, E2 extends Exception> void ifNotNullOrElse(Try.Consumer<? super T, E> action, Try.Runnable<E2> emptyAction) throws E, E2 {
         N.checkArgNotNull(action);